requirements.txt
customtkinter>=5.2.0
PyPDF2>=3.0.0
Pillow>=9.0.0 